% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/me_densFromMoments.R
\name{me_densFromMoments}
\alias{me_densFromMoments}
\title{probability density function from distribution moments using maximum entropy}
\usage{
me_densFromMoments(mu, x, xEval = NULL, lambda = NULL, eps = 1e-05,
  verbose = F)
}
\arguments{
\item{mu}{a vector containing moments}

\item{x}{the vector where to evaluate the me probability density function}

\item{lambda}{(optional) a vector for starting values of Lagrangian parameters. If the value of lambda is not provided by the user, a uniform distribution is used.}

\item{eps}{a number corresponding to the tolerance for the stopping rule. 
The iterations stop when the relative error is less that eps.}
}
\value{
a list containing 
\enumerate{
\item a vector of Lagrangian parameters,
\item a vector containing values of the pdf at locations indicated in input parameter \code{x}
\item a vector containing values for entropy calculated at each iteration
}
}
\description{
\code{me_densFromMoments} calculates the probability density function corresponding to moments using the principle of maximum entropy
}
\details{
Some more detailed description. Authors : Karina Cucchi, Changhong Wang
}
\examples{
x=seq(-6,6,by=0.1)
mu = c(0.0000,5.4300,2.6779,51.3181,50.0401)
output = me_densFromMoments(mu = mu,x = x,eps = 1e-30)
lambda=output[[1]]
print(paste0('lambda = [',paste(round(lambda,3),collapse = ' '),']'))

}

