% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/me_densFromSamples.R
\name{me_densFromSamples}
\alias{me_densFromSamples}
\title{probability density function evaluation from distribution samples using maximum entropy}
\usage{
me_densFromSamples(q, nbMoments, x = NULL, xEval = NULL, lambda = NULL,
  eps = 1e-05, verbose = F)
}
\arguments{
\item{q}{the vector containing samples from the distribution}

\item{nbMoments}{an integer corresponding to the number of moments to account for in the calculation of the me pdf}

\item{x}{is a vector containing values to evaluate the pdf and necessary to calculate the entropy.
If x is not defined, the default is seq(min(q)-sd(q),max(q)+sd(q),length=1000)}

\item{xEval}{vector containing where to evaluate the probability density function.
If xEval is not specified, it is set to x.}

\item{eps}{a number corresponding to the tolerance for the stopping rule.}

\item{verbose}{a boolean indicating whether to allow for printing outputs
The iterations stop when the relative error is less that eps.}
}
\value{
a list containing 
\enumerate{
\item a vector of Lagrangian parameters,
\item a vector containing values of the pdf at locations indicated in input parameter \code{x}
\item a vector containing values for entropy calculated at each iteration
}
}
\description{
\code{me_densFromSamples} calculates the probability density function corresponding to distribution samples using the principle of maximum entropy
}
\details{
Some more detailed description. Authors : Karina Cucchi, Changhong Wang
}
\examples{
x=seq(-6,6,by=0.1)
q=c(3.132,3.132,3.132,-1.625,-1.821,1.026,-3.003,-2.332,
0.827,-0.437,0.305,-3.433,1.865,-1.131,-2.828,3.921,-0.730)
output = me_densFromSamples(q=q,nbMoments=5,x = x,eps = 1e-30)
lambda=output[[1]]
print(paste0('lambda = [',paste(round(lambda,3),collapse = ' '),']'))
histData = hist(x = q,freq = F,xlim=range(x),col = 'lightblue')
lines(x,p,col='red')

}

